Backend bundle — make the LLM backend swappable
Three small refactors that turn the implicit-Claude-default-but-
injectable shape into a registered-backends shape. No behaviour
change for existing callers.

New: src/f3dasm/_src/agentic/backends/

  __init__.py    — empty subpackage marker
  base.py        — Backend frozen dataclass + StrategizerSession,
                   ImplementerSession Protocols (moved from
                   agent_runtime.py so backends can implement
                   against them without importing the runtime)
  claude.py      — _ClaudeStrategizer, _ClaudeImplementer,
                   _classify_sdk_error, _preflight_claude_cli,
                   _strategizer_factory, _implementer_factory,
                   plus CLAUDE_BACKEND = Backend(name="claude",
                   default_model="claude-haiku-4-5-20251001", …)

agent_runtime.py:
  - Imports StrategizerSession/ImplementerSession from backends.base
  - Imports CLAUDE_BACKEND from backends.claude
  - AgenticRun.__init__ gains a `backend: Backend | None = None`
    kwarg, defaulting to CLAUDE_BACKEND. The existing
    strategizer_factory / implementer_factory kwargs remain as
    test-injection overrides; when not supplied they fall through
    to backend.strategizer_factory / .implementer_factory.
  - `model` kwarg default changes from MVP_DEFAULT_MODEL to
    None; resolves to backend.default_model when not given.
  - _preflight_check_claude_cli replaced by a one-liner calling
    self._backend.preflight() (skipped when stub factories are
    injected; existing skip-on-test behaviour preserved).
  - Removed: the _ClaudeStrategizer/Implementer classes, the
    _classify_sdk_error helper, the old _default_*_factory pair,
    and an unused `import asyncio`.
  - MVP_DEFAULT_MODEL still importable from agent_runtime for
    backward-compat (= CLAUDE_BACKEND.default_model).

Public surface: `from f3dasm.agentic import Backend, CLAUDE_BACKEND`
now works alongside the existing exports.

Tests: 138 prior pass; one new (test_custom_backend_drops_in_via_kwarg)
constructs a stub Backend and verifies the kwarg path threads through
correctly. 139 total. Ruff clean. Test-import path for
_classify_sdk_error updated to its new home.

Sized for swap-ability: adding an OpenAI or Gemini backend now
requires only a new file under backends/, providing matching
strategizer_factory / implementer_factory / preflight / error
classification, plus a registered Backend instance.
